fix(artifact_source): move the module under IO to avoid circular import
