+++ queenbee/io/artifact_source.py
